requirements-eeg.txt
python-socketio[asyncio-client]>=5.11.0
aiohttp>=3.9.0

